修改限流的规则 6v以下3.5A 10v一下限制20w 10v以上限制18w 修复电流显示超过2.56A后溢出的bug

+++ src/power.h
@@ -1,7 +1,7 @@
 /*
  * @Author: your name
  * @Date: 2021-04-18 09:32:57
- * @LastEditTime: 2021-05-14 01:03:22
+ * @LastEditTime: 2021-05-24 22:40:34
  * @LastEditors: huzhenhong
  * @Description: In User Settings Edit
  * @FilePath: \WSJ\src\power.h
@@ -12,8 +12,8 @@
 void stop8812(void);
 void init8812(void);
 void startPow(void);
-u8 GetIBus();
-u8 GetIBusAvg();
+u16 GetIBus();
+u16 GetIBusAvg();
 u8 GetVBatAvg();
 //u8 GetIBatAvg();
 void SetVolt(u8 v);
